Load block icons SVG's using smart textures in non-ideal way.
This approach is flawed because it requires an image download after the loading
phase when textures were loaded. It is also flawed because drawing SVG's on the
canvas in this way does not allow any control over the draw color of the SVG.

Ideally, need to find a way to load the SVG XML data into the client, and write
some code to parse/modify the SVG data before drawing to the canvas.