Add ThrusterBlock class and add a ThrusterBlock to the player's ship.
Contributes to LEO-226 and LEO-228. The ThrusterBlocks are not quite positioned
correctly because the BlockGrid class currently has no way of describing empty
spaces so arbitrary positioning of blocks is not yet supported.

ThrusterBlocks also do not yet have an icon because no icon was included for
them in the assets folder.