Change behavior of Deployment selector to not append :2000 to all URLs, only to local or unknown URLs.
If we hardcoded port 2000, this would not work with the Azure port environment variable. Azure provides each Node instance with a named pipe that effectively routes to http://hostname. In this case, the server will be routed to http://tl-cosmos-server-<stage>.azurewebsites.net, which is aliased to http://<stage>.srv.cosmos.teamleonine.com.