Fix issue where a new joining player could somewhat control the player that joined immediately before it.
The bug is reproducible by having two or more players A and B. B joins after A. When B presses the "up" or "down" arrow keys to apply forwards or backwards impulse, the impulse is instead applied to A's ship in relation to B's orientation.
This means that if B is oriented differently from A, A will spin uncontrollably in addition to moving in the direction that B is oriented.